VB-3827 - Rename enum to RestrictionType to better describe enum
