+++ src/test/kotlin/uk/gov/justice/digital/hmpps/prisonercontactregistry/integration/IntegrationTestBase.kt
@@ -15,7 +15,7 @@ import uk.gov.justice.digital.hmpps.prisonercontactregistry.dto.AddressDto
 import uk.gov.justice.digital.hmpps.prisonercontactregistry.dto.ContactDto
 import uk.gov.justice.digital.hmpps.prisonercontactregistry.dto.ContactsDto
 import uk.gov.justice.digital.hmpps.prisonercontactregistry.dto.RestrictionDto
-import uk.gov.justice.digital.hmpps.prisonercontactregistry.enum.Restriction
+import uk.gov.justice.digital.hmpps.prisonercontactregistry.enum.RestrictionType
 import uk.gov.justice.digital.hmpps.prisonercontactregistry.helper.JwtAuthHelper
 import uk.gov.justice.digital.hmpps.prisonercontactregistry.integration.mock.PrisonApiMockServer
 import java.time.LocalDate
@@ -81,7 +81,7 @@ abstract class IntegrationTestBase {
     assertThat(contact.nextOfKin).isFalse
     assertThat(contact.commentText).isEqualTo("Comment Here")
     assertThat(contact.restrictions.size).isEqualTo(1)
-    assertThat(contact.restrictions[0].restrictionType).isEqualTo(Restriction.BANNED.toString())
+    assertThat(contact.restrictions[0].restrictionType).isEqualTo(RestrictionType.BANNED.toString())
     assertThat(contact.restrictions[0].restrictionTypeDescription).isEqualTo("Banned")
     assertThat(contact.restrictions[0].startDate).isEqualTo("2012-09-13")
     assertThat(contact.restrictions[0].expiryDate).isEqualTo("2014-09-13")
@@ -115,7 +115,7 @@ abstract class IntegrationTestBase {
     assertThat(contact.emergencyContact).isFalse
     assertThat(contact.nextOfKin).isFalse
     assertThat(contact.restrictions.size).isEqualTo(1)
-    assertThat(contact.restrictions[0].restrictionType).isEqualTo(Restriction.BANNED.toString())
+    assertThat(contact.restrictions[0].restrictionType).isEqualTo(RestrictionType.BANNED.toString())
     assertThat(contact.restrictions[0].restrictionTypeDescription).isEqualTo("Banned")
     assertThat(contact.restrictions[0].startDate).isEqualTo("2012-09-13")
     assertThat(contact.restrictions[0].globalRestriction).isEqualTo(false)
@@ -184,7 +184,7 @@ abstract class IntegrationTestBase {
     comment: String = "Comment Here",
   ): RestrictionDto {
     return RestrictionDto(
-      restrictionType = Restriction.BANNED.toString(),
+      restrictionType = RestrictionType.BANNED.toString(),
       restrictionTypeDescription = "Banned",
       startDate = startDate,
       expiryDate = expiryDate,
